Started adding ending sequence stuff

+++ src/controllers/GameScreenController.java
@@ -41,7 +41,7 @@ import java.io.IOException;
         private Character character;
         ArrayList<Dialog> dialogslist = new ArrayList<Dialog>();
         private int dialogNum = 0;
-        static int numCharactersVisited = 0;
+        static double numCharactersVisited = 0;
